refactor: заменить лайки на оценки

+++ src/main/java/ru/yandex/practicum/filmorate/model/data/FilmEntity.java
@@ -13,7 +13,7 @@ public class FilmEntity {   // Памяти и немного замедляет
     String description;
     LocalDate releaseDate;
     int duration;
-    byte rate;
+    float rate;
     int rating;
 
 }
